chore: Update user and todo types and models

+++ src/controllers/todo.controller.ts
@@ -1,5 +1,12 @@
 import { Request, Response } from 'express'
-import { addTodoToDB, getTodosFromDB, updateTodoById, deleteTodoById } from '../services/todo.service'
+import {
+  addTodoToDB,
+  getTodosFromDB,
+  updateTodoById,
+  deleteTodoById,
+  getTodoById,
+  deleteSelectedTodos
+} from '../services/todo.service'
 import { v4 as uuidv4 } from 'uuid'
 import { createTodoValidation, updateTodoValidation } from '../validations/todo.validation'
 
@@ -28,8 +35,11 @@ export const createTodo = async (req: Request, res: Response) => {
 }
 
 export const getTodos = async (req: Request, res: Response) => {
+  const userId: string = res.locals.user.id
+  const { search } = req.query
+
   try {
-    const todos = await getTodosFromDB()
+    const todos = await getTodosFromDB(userId, search?.toString())
     console.info('Get todos success')
     return res.status(200).send({
       status: true,
@@ -44,9 +54,16 @@ export const getTodos = async (req: Request, res: Response) => {
 }
 
 export const updateTodo = async (req: Request, res: Response) => {
+  const userId: string = res.locals.user.id
   const {
     params: { id }
   } = req
+
+  const todo = await getTodoById(id)
+
+  if (todo?.user_id !== userId) {
+    return res.status(403).send({ status: false, statusCode: 403, message: 'Forbidden' })
+  }
 
   const { error, value } = updateTodoValidation(req.body)
 
@@ -76,26 +93,64 @@ export const updateTodo = async (req: Request, res: Response) => {
 }
 
 export const deleteTodo = async (req: Request, res: Response) => {
+  const userId: string = res.locals.user.id
   const {
-    params: { id }
+    params: { id },
+    body: { ids }
   } = req
 
-  try {
-    const result = await deleteTodoById(id)
+  if (ids) {
+    const result = await Promise.all(
+      ids.map(async (id: string) => {
+        const todo = await getTodoById(id)
 
-    if (result) {
-      console.log('Delete todo success')
-      return res.status(200).send({
-        status: true,
-        statusCode: 200,
-        message: 'Delete todo success'
+        if (todo?.user_id !== userId) {
+          return false
+        }
       })
-    } else {
+    )
+
+    if (result.includes(false)) {
       console.log('Data not found')
       return res.status(404).send({ status: false, statusCode: 404, message: 'Data not found' })
     }
-  } catch (error) {
-    console.error('ERR: todo - delete = ', error)
-    return res.status(422).send({ status: false, statusCode: 422, message: error })
+
+    try {
+      await deleteSelectedTodos(ids)
+      console.log('Delete todos success')
+      return res.status(200).send({
+        status: true,
+        statusCode: 200,
+        message: 'Delete todos success'
+      })
+    } catch (error) {
+      console.error('ERR: todo - delete = ', error)
+      return res.status(422).send({ status: false, statusCode: 422, message: error })
+    }
+  } else {
+    const todo = await getTodoById(id)
+
+    if (todo?.user_id !== userId) {
+      return res.status(403).send({ status: false, statusCode: 403, message: 'Forbidden' })
+    }
+
+    try {
+      const result = await deleteTodoById(id)
+
+      if (result) {
+        console.log('Delete todo success')
+        return res.status(200).send({
+          status: true,
+          statusCode: 200,
+          message: 'Delete todo success'
+        })
+      } else {
+        console.log('Data not found')
+        return res.status(404).send({ status: false, statusCode: 404, message: 'Data not found' })
+      }
+    } catch (error) {
+      console.error('ERR: todo - delete = ', error)
+      return res.status(422).send({ status: false, statusCode: 422, message: error })
+    }
   }
 }
